requirements.txt
autopep8==2.0.1
beautifulsoup4==4.11.2
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==3.0.1
idna==3.4
isort==5.12.0
psycopg2-binary==2.9.5
pycodestyle==2.10.0
python-dotenv==0.21.1
requests==2.28.2
soupsieve==2.3.2.post1
tomli==2.0.1
urllib3==1.26.14
